requirements.txt
flask
replicate
python-dotenv
gunicorn